fix FamilyData model to load properly from json
