+++ Program.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using System.Text.Json;
 using src;
+using IRS;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
 builder.RootComponents.Add<App>("#app");
@@ -11,10 +12,14 @@ var httpClient = new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
 builder.Services.AddScoped<ISearchModel>(sp => SearchModel.Create());
 
-FamilyYears? familyYears = await FamilyYears.Create(httpClient);
-if (familyYears != null) {
-    builder.Services.AddSingleton<IFamilyYears>(familyYears);
+IRSData? irsData = await IRSData.Create(httpClient);
+if (irsData != null) {
+    builder.Services.AddSingleton<IRSData>(irsData);
+    builder.Services.AddSingleton<IFamilyData>(new FamilyData(irsData));
+} else {
+    throw new Exception("irsData is null");
 }
+
 
 var fundsJson = await httpClient.GetAsync("https://raw.githubusercontent.com/bogle-tools/site/main/wwwroot/data/funds.json");
 var Funds = await JsonSerializer.DeserializeAsync<List<Fund>>(fundsJson.Content.ReadAsStream());
